+++ CHANGELOG.markdown
@@ -3,6 +3,8 @@
 * Added some missing examples/ files to the distribution.
 * Renamed `Data.Bits.Lens.traverseBits` to `bits`.
 * Removed `(^!?)`, which was an alias for `(^?!)`.
+* Removed the need for `Trustworthy` by changing the implementation of `coerce` for `BazaarT`.
+* Moved BazaarT to `Control.Lens.Internal`.
 
 3.5.1
 -----